filtering comment findings now, still not editing existing comment
