+++ lunatrace/bsl/backend/src/models/scan.ts
@@ -21,18 +21,25 @@ import {
   Findings_Constraint,
   Findings_Insert_Input,
   Findings_Update_Column,
+  InsertScanMutation,
   Scans_Insert_Input,
 } from '../hasura-api/generated';
 import { Convert, GrypeScanReport, Match } from '../types/grype-scan-report';
 
-export async function parseAndUploadScan(sbomStream: Readable, buildId: string): Promise<Scans_Insert_Input> {
+export type InsertedScan = NonNullable<InsertScanMutation['insert_scans_one']>;
+
+export async function parseAndUploadScan(sbomStream: Readable, buildId: string): Promise<InsertedScan> {
   const rawGrypeReport = await runGrypeScan(sbomStream);
   const typedRawGrypeReport = Convert.toScanReport(rawGrypeReport);
   const scan = await parseScan(typedRawGrypeReport, buildId);
-  await hasura.InsertScan({
+  const insertRes = await hasura.InsertScan({
     scan,
+    build_id: buildId,
   });
-  return scan;
+  if (!insertRes.insert_scans_one) {
+    throw new Error(`Failed to insert scan into hasura, resp: ${JSON.stringify(insertRes)}`);
+  }
+  return insertRes.insert_scans_one;
 }
 
 export async function runGrypeScan(sbomStream: Readable): Promise<string> {
