+++ lunatrace/bsl/backend/src/sqs-handlers/scan-sbom.ts
@@ -14,21 +14,16 @@
 import { Readable } from 'stream';
 import zlib from 'zlib';
 
-import markdownTable from 'markdown-table';
 import validate from 'validator';
 
-import { generateGithubGraphqlClient } from '../github';
-import { getInstallationAccessToken } from '../github/auth';
+import { commentOnPrIfExists } from '../github/pr-comment-generator';
 import { hasura } from '../hasura-api';
-import { Scans_Insert_Input } from '../hasura-api/generated';
-import { parseAndUploadScan } from '../models/scan';
+import { InsertedScan, parseAndUploadScan } from '../models/scan';
 import { S3ObjectMetadata } from '../types/s3';
 import { SbomBucketInfo } from '../types/scan';
 import { QueueErrorResult, QueueSuccessResult } from '../types/sqs';
 import { aws } from '../utils/aws-utils';
 import { isError, tryF } from '../utils/try';
-
-import { groupByPackage, VulnerablePackage } from './report-generator/group-findings';
 
 function decompressGzip(stream: Readable, streamLength: number): Promise<zlib.Gzip> {
   return new Promise((resolve, reject) => {
@@ -50,7 +45,7 @@ function decompressGzip(stream: Readable, streamLength: number): Promise<zlib.Gz
   });
 }
 
-async function scanSbom(buildId: string, sbomBucketInfo: SbomBucketInfo): Promise<Scans_Insert_Input> {
+async function scanSbom(buildId: string, sbomBucketInfo: SbomBucketInfo): Promise<InsertedScan> {
   console.log(`[buildId: ${buildId}]`, `scanning sbom ${JSON.stringify(sbomBucketInfo)}`);
 
   const [sbomStream, sbomLength] = await aws.getFileFromS3(
@@ -74,102 +69,6 @@ async function scanSbom(buildId: string, sbomBucketInfo: SbomBucketInfo): Promis
   return scanReport;
 }
 
-function formatLocationText(finding: VulnerablePackage) {
-  if (finding.locations.length === 0) {
-    return 'Unknown';
-  }
-
-  return `${finding.locations.length} location${finding.locations.length > 1 ? 's' : ''}`;
-}
-
-function generatePullRequestCommentFromReport(projectId: string, report: Scans_Insert_Input) {
-  const messageParts = [
-    '## Build Snapshot Complete',
-    '',
-    `\n[View Full Report](https://lunatrace.lunasec.io/project/${projectId}/build/${report.build_id})`,
-  ];
-
-  if (!report.findings) {
-    return messageParts.join('\n');
-  }
-
-  const groupedFindings = groupByPackage(projectId, report.findings.data);
-
-  const filteredFindings = Object.values(groupedFindings).filter(
-    // TODO: Make the severity filter configurable.
-    (finding) => finding.severity !== 'Unknown' && finding.severity === 'Critical'
-  );
-
-  const tableData: string[][] = filteredFindings.map((finding): string[] => {
-    return [
-      finding.package_name,
-      finding.package_versions ? finding.package_versions.join(', ') : 'Unknown Version',
-      finding.severity,
-      formatLocationText(finding),
-      `[Dismiss](https://lunatrace.lunasec.io/project/${projectId}/build/${report.build_id})`,
-    ] as string[];
-  });
-
-  messageParts.push('### Security Scan Findings');
-  messageParts.push(`Showing ${tableData.length} results.\n`);
-  messageParts.push(markdownTable([['Package Name', 'Versions', 'Severity', 'Locations', ''], ...tableData]));
-
-  return messageParts.join('\n');
-}
-
-async function commentOnPrIfExists(buildId: string, scanReport: Scans_Insert_Input) {
-  const buildLookup = await hasura.GetScanReportNotifyInfoForBuild({
-    build_id: buildId,
-  });
-
-  if (
-    !buildLookup.builds_by_pk ||
-    !buildLookup.builds_by_pk.project ||
-    !buildLookup.builds_by_pk.project.organization
-  ) {
-    console.error(`unable to get required scan notify information for buildId: ${buildId}`);
-    return;
-  }
-
-  const projectId = buildLookup.builds_by_pk.project.id;
-  const installationId = buildLookup.builds_by_pk.project.organization.installation_id;
-  const pullRequestId = buildLookup.builds_by_pk.pull_request_id;
-
-  if (!installationId) {
-    console.log(
-      `installation id is not defined for the organization linked to build: ${buildId}, skipping github PR comment`
-    );
-    return;
-  }
-
-  if (!pullRequestId) {
-    console.log(
-      `pull request id is not defined for buildId: ${buildId}, skipping comment because this build did not come from a PR`
-    );
-    return;
-  }
-
-  const installationToken = await getInstallationAccessToken(installationId);
-
-  const github = generateGithubGraphqlClient(installationToken);
-
-  const body = generatePullRequestCommentFromReport(projectId, scanReport);
-
-  if (body === null) {
-    console.error(`generated scan report is null`, {
-      projectId,
-      installationId,
-      pullRequestId,
-    });
-    return;
-  }
-
-  await github.AddComment({
-    subjectId: pullRequestId.toString(),
-    body,
-  });
-}
-
 export async function handleScanSbom(message: S3ObjectMetadata): Promise<QueueSuccessResult | QueueErrorResult> {
   const { key, region, bucketName } = message;
   const buildId = key.split('/').pop();
@@ -184,7 +83,7 @@ export async function handleScanSbom(message: S3ObjectMetadata): Promise<QueueSu
 
   const bucketInfo: SbomBucketInfo = { region, bucketName, key };
 
-  const scanResp = await tryF<Scans_Insert_Input>(async () => await scanSbom(buildId, bucketInfo));
+  const scanResp = await tryF(async () => await scanSbom(buildId, bucketInfo));
 
   if (isError(scanResp)) {
     console.error('Sbom Scanning Error:', { scanResp });
