+++ lunatrace/bsl/frontend/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "dependencies": {
     "@aws-sdk/client-s3": "^3.52.0",
-    "@lunatrace/lunatrace-common": "link:../common",
+    "@lunatrace/lunatrace-common": "workspace:~",
     "@ory/integrations": "~0.2.2",
     "@ory/kratos-client": "0.9.0-alpha.3",
     "@ory/themes": "~0.0.101",
