+++ lunatrace/bsl/hasura/metadata/databases/lunatrace/tables/public_ignored_vulnerabilities.yaml
@@ -32,6 +32,16 @@ insert_permissions:
 select_permissions:
 - permission:
     columns:
+    - locations
+    - note
+    - creator_id
+    - id
+    - project_id
+    - vulnerability_id
+    filter: {}
+  role: service
+- permission:
+    columns:
     - id
     - project_id
     - vulnerability_id
