+++ lunatrace/bsl/hasura/metadata/databases/lunatrace/tables/public_organizations.yaml
@@ -41,6 +41,7 @@ insert_permissions:
   role: user
 select_permissions:
 - permission:
+    allow_aggregations: true
     columns:
     - name
     - github_id
